+++ test/domain-shared/env.test.js
@@ -1,0 +1,5 @@
+describe("NODE_ENV", () => {
+    it("environment should be test", () => {
+        expect(process.env.NODE_ENV).toBe("test");
+    });
+})